feat: enforce tailwind color tokens via eslint
Add custom eslint plugin to ban arbitrary hex colors in className.
Update lint script with --max-warnings 0.
